Upgrade project to Java 17 and Gradle 8

+++ modules/synth/src/main/kotlin/com/bartlomiejpluta/smnp/ext/synth/lib/envelope/Envelope.kt
@@ -10,7 +10,7 @@ abstract class Envelope {
          (byte * eval(
             index.toDouble(),
             wave.bytes.size
-         )).toByte()
+         )).toInt().toByte()
       }.toByteArray())
    }
